app/requirements.txt
flask   
flask-restx
flask-sqlalchemy
python-dotenv
openai
aiohttp==3.9.0b
pymysql
python-decouple
Flask-Cors
mysqlclient
python-dotenv